refactor: move Toast and ToastType out of view layer
