+++ src/main/java/edu/ntnu/idatt2003/millions/service/toast/Toast.java
@@ -1,3 +1,3 @@
-package edu.ntnu.idatt2003.millions.view.component.toast;
+package edu.ntnu.idatt2003.millions.service.toast;
 
 public record Toast(long id, ToastType type, String message) {}
