+++ src/main/java/edu/ntnu/idatt2003/millions/service/toast/ToastType.java
@@ -1,0 +1,6 @@
+package edu.ntnu.idatt2003.millions.service.toast;
+
+public enum ToastType {
+    SUCCESS,
+    ERROR
+}
