--- src/main/java/edu/ntnu/idatt2003/millions/view/component/toast/ToastType.java
@@ -1,6 +1,0 @@
-package edu.ntnu.idatt2003.millions.view.component.toast;
-
-public enum ToastType {
-    SUCCESS,
-    ERROR
-}
